Pikmin now follow the bridge as it's being attacked
NOTE: I need to rethink how I'm going to handle all of these different states and these variables, the code is getting messy and as said in the specifications that's something we DONT want to do. My immediate thoughts are to strip the whole thing down to something that moves and slowly build it back up again, will probably do tommorow though.